# public/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OM Paye API",
        "description": "API pour OM Paye - Système de paiement mobile",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Créer un nouveau compte (Admin uniquement)",
                "operationId": "c7bb21a0d87049e161b8e4c9b0d909f2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prenom",
                                    "cni",
                                    "telephone",
                                    "type"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Dupont"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Jean"
                                    },
                                    "cni": {
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "782345678"
                                    },
                                    "sexe": {
                                        "type": "string",
                                        "enum": [
                                            "M",
                                            "F"
                                        ],
                                        "example": "M"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "marchand",
                                            "utilisateur"
                                        ],
                                        "example": "utilisateur"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "motdepasse123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compte créé avec succès"
                    },
                    "403": {
                        "description": "Accès non autorisé"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/send-otp": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Envoyer un code OTP par SMS",
                "operationId": "bffceebe22269794653d542fb5e548dd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "telephone"
                                ],
                                "properties": {
                                    "telephone": {
                                        "type": "string",
                                        "example": "782345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP envoyé"
                    },
                    "400": {
                        "description": "Numéro invalide"
                    }
                }
            }
        },
        "/api/auth/verify-otp": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Vérifier le code OTP et obtenir un token",
                "operationId": "abcb5741589707b15c7ecac7a67c13aa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "telephone",
                                    "code"
                                ],
                                "properties": {
                                    "telephone": {
                                        "type": "string",
                                        "example": "782345678"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token généré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string"
                                        },
                                        "user": {
                                            "type": "object"
                                        },
                                        "expires_at": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Code invalide"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Connexion avec téléphone et mot de passe",
                "operationId": "dd51a484b7fa0f97c7680fdbde14cf5f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "telephone",
                                    "password"
                                ],
                                "properties": {
                                    "telephone": {
                                        "type": "string",
                                        "example": "782345678"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "motdepasse123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Identifiants invalides"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Déconnexion",
                "operationId": "df9bdc6c7a6d9f20b983526b2ec0f698",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie"
                    },
                    "401": {
                        "description": "Token invalide"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Obtenir les informations de l'utilisateur connecté",
                "operationId": "b5de204f5e452ef51327c691aa99e77c",
                "responses": {
                    "200": {
                        "description": "Informations utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token invalide"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token JWT pour l'authentification",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}